+++ javascript/node/selenium-webdriver/lib/command.js
@@ -161,7 +161,6 @@ const Name = {
   DISMISS_ALERT: 'dismissAlert',
   GET_ALERT_TEXT: 'getAlertText',
   SET_ALERT_TEXT: 'setAlertValue',
-  SET_ALERT_CREDENTIALS: 'setAlertCredentials',
 
   GET_AVAILABLE_LOG_TYPES: 'getAvailableLogTypes',
   GET_LOG: 'getLog',