+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -120,6 +120,8 @@ mode.
        -  InvalidElementCoordinatesError
 *  Changes to `lib/webdriver.WebDriver`:
    -  Dropped support for "requiredCapabilities" from WebDriver.createSession
+*  Changes to `lib/webdriver.Alert`:
+   -  Removed authenticateAs
 *  Changes to `lib/webdriver.Options` (`driver.manage()`):
    -  Removed timeouts (use get/setTimeouts)
 *  Changes to `lib/webdriver.Window` (`driver.manage().window()`):